+++ bin/read.js
@@ -10,12 +10,16 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 var axios_1 = __importDefault(require("axios"));
 var js_yaml_1 = __importDefault(require("js-yaml"));
 var fs_1 = __importDefault(require("fs"));
 var path_1 = __importDefault(require("path"));
 var T = __importStar(require("fp-ts/lib/Task"));
 var E = __importStar(require("fp-ts/lib/Either"));
+var pipeable_1 = require("fp-ts/lib/pipeable");
+var utils_1 = require("./utils");
+var function_1 = require("fp-ts/lib/function");
 function isUrl(pathOrUrl) {
     return pathOrUrl.indexOf("https://") !== -1;
 }
@@ -48,8 +52,20 @@ function getContentFromPath(file) {
     };
 }
 function getContent(source) {
-    return isUrl(source)
-        ? getContentFromURL(source)
-        : T.fromIO(getContentFromPath(source));
+    return pipeable_1.pipe(source, utils_1.doIf(isUrl, function (source) { return getContentFromURL(source); }, function (source) { return T.fromIO(getContentFromPath(source)); }));
 }
-exports.getContent = getContent;
+function patchSwagger(patchSource) {
+    return function (swagger) {
+        return pipeable_1.pipe(patchSource, function (source) { return getContent(source); }, T.map(E.map(utils_1.patch(swagger))));
+    };
+}
+function getSwagger(patchSource) {
+    return function (source) {
+        return pipeable_1.pipe(source, function (source) { return getContent(source); }, utils_1.doIf(function () { return patchSource != null; }, function (task) {
+            return pipeable_1.pipe(task, T.chain(function (eitherSwagger) {
+                return E.either.traverse(T.task)(eitherSwagger, patchSwagger(patchSource));
+            }), T.map(E.flatten));
+        }, function_1.identity));
+    };
+}
+exports.getSwagger = getSwagger;